Shift props and handlers into MainGraph Component
