+++ src/components/MainContainer.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { Layout } from 'antd';
-import _ from 'lodash';
+import { margin } from '../utils/constants';
 
 import Search from './Search/Search'
 import MainGraph from './Graph/MainGraph';
@@ -8,60 +8,20 @@ import MainChart from './Chart/MainChart';
 import MainMap from './Map/MainMap';
 import Methodology from './Methodology';
 
-import { instantiateScenarios, getRange } from '../utils/utils'
-import { utcParse, timeFormat } from 'd3-time-format'
-import { timeDay } from 'd3-time'
-import { maxIndex } from 'd3-array';
-import { STATS, LEVELS, margin } from '../utils/constants';
-
 const dataset = require('../store/geo06085.json');
-const geojsonStats = require('../store/statsForMap.json')
 
-const parseDate = utcParse('%Y-%m-%d')
-const formatDate = timeFormat('%Y-%m-%d')
 
 class MainContainer extends Component {
     constructor(props) {
         super(props);
         this.state = {
             dataset: {},
-            dataLoaded: false,
-            series: {},
-            seriesList: [{}],
-            allTimeSeries: {},
-            dates: [],
-            allTimeDates: [],
-            yAxisLabel: '',
-            stat: STATS[0],
+            dataLoaded: false, // TODO: might not need this here!
             geoid: '06085', 
-            SCENARIOS: [],
-            scenario: {},
-            scenarioList: [],           
-            severity: _.cloneDeep(LEVELS[0]), 
-            severityList: [_.cloneDeep(LEVELS[0])],
-            scenarioHovered: '',
-            statThreshold: 0,
-            statSliderActive: false,
-            seriesMax: Number.NEGATIVE_INFINITY,
-            seriesMin: Number.POSITIVE_INFINITY,
-            dateThreshold: new Date(),
-            dateRange: [parseDate('2020-03-01'), parseDate('2020-09-01')],
-            firstDate: '',
-            lastDate: '',
-            r0: [0, 4],
-            simNum: '150',
-            percExceedenceList: [],
-            showConfBounds: false,
-            confBounds: {},
-            confBoundsList: [{}],
-            showActual: false,
             graphW: 0,
             graphH: 0,
-            brushActive: false,
-            animateTransition: true,
-            scenarioClickCounter: 0,
             mapContainerW: 0,
-            mapContainerH: 0,
+            mapContainerH: 0
         };
     };
 
@@ -69,86 +29,70 @@ class MainContainer extends Component {
         // console.log('componentDidMount')
         console.log('dataset', dataset)
         
-        // shift to Children or keep in MainContainer?
+        // TODO: shift to Children or keep in MainContainer?
         window.addEventListener('resize', this.updateGraphDimensions)
         window.addEventListener('resize', this.updateMapContainerDimensions)
         this.updateGraphDimensions()
         this.updateMapContainerDimensions()
         
-        const [SCENARIOS, scenario, scenarioList, scenarioListChart, scenarioMap] =
-            instantiateScenarios(dataset);
+        // const [SCENARIOS, scenario, scenarioList, scenarioListChart, scenarioMap] =
+        //     instantiateScenarios(dataset);
 
-        // instantiate initial series and dates
-        const { severity, stat } = this.state;
-        const series = dataset[scenario.key][severity.key][stat.key].sims;
+        // // instantiate initial series and dates
+        // const { severity, stat } = this.state;
+        // const series = dataset[scenario.key][severity.key][stat.key].sims;
          
-        const dates = dataset[scenario.key].dates.map( d => parseDate(d));
-        const firstDate = dates[0];
-        const lastDate = dates[dates.length - 1];
+        // const dates = dataset[scenario.key].dates.map( d => parseDate(d));
+        // const firstDate = dates[0];
+        // const lastDate = dates[dates.length - 1];
         
-        const seriesPeaks = series.map(sim => sim.max);
-        const [seriesMin, seriesMax] = getRange(seriesPeaks);
-        const statThreshold = Math.ceil((seriesMax / 1.4) / 100) * 100;
+        // const seriesPeaks = series.map(sim => sim.max);
+        // const [seriesMin, seriesMax] = getRange(seriesPeaks);
+        // const statThreshold = Math.ceil((seriesMax / 1.4) / 100) * 100;
 
-        // iterate through SeriesList
-        const simsOver = this.updateThresholdIterate(
-            series,
-            statThreshold,
-            dates,
-            this.state.dateThreshold
-            )        
-        const percExceedence = simsOver / series.length;
+        // // iterate through SeriesList
+        // const simsOver = this.updateThresholdIterate(
+        //     series,
+        //     statThreshold,
+        //     dates,
+        //     this.state.dateThreshold
+        //     )        
+        // const percExceedence = simsOver / series.length;
 
-        // send these sims to brush after it's been colored
-        // include in SeriesList loop
-        // if (firstIndex of SeriesList)
-        const allTimeSeries = Array.from(series)
-        // shouldn't happen twice
-        const allTimeDates = Array.from(dates)
+        // // send these sims to brush after it's been colored
+        // // include in SeriesList loop
+        // // if (firstIndex of SeriesList)
+        // const allTimeSeries = Array.from(series)
+        // // shouldn't happen twice
+        // const allTimeDates = Array.from(dates)
 
-        // take out of loop so not redundant
-        const idxMin = timeDay.count(firstDate, this.state.dateRange[0]);
-        const idxMax = timeDay.count(firstDate, this.state.dateRange[1]);
-        const newDates = Array.from(allTimeDates.slice(idxMin, idxMax));
-        const filteredSeries = series.map( s => {
-            const newS = {...s}
-            newS.vals = s.vals.slice(idxMin, idxMax)
-            return newS
-        })
+        // // take out of loop so not redundant
+        // const idxMin = timeDay.count(firstDate, this.state.dateRange[0]);
+        // const idxMax = timeDay.count(firstDate, this.state.dateRange[1]);
+        // const newDates = Array.from(allTimeDates.slice(idxMin, idxMax));
+        // const filteredSeries = series.map( s => {
+        //     const newS = {...s}
+        //     newS.vals = s.vals.slice(idxMin, idxMax)
+        //     return newS
+        // })
         
-        const yAxisLabel = `Daily ${stat.name}`;
-        const percExceedenceList = [percExceedence]
+        // const yAxisLabel = `Daily ${stat.name}`;
+        // const percExceedenceList = [percExceedence]
 
-        // instantiate confidence bounds
-        const confBounds = dataset[scenario.key][severity.key][stat.key].conf;
-        const filteredConfBounds = confBounds.slice(idxMin, idxMax)
+        // // instantiate confidence bounds
+        // const confBounds = dataset[scenario.key][severity.key][stat.key].conf;
+        // const filteredConfBounds = confBounds.slice(idxMin, idxMax)
 
-        // instantiate start and end date (past 2 weeks) for summary stats
-        const summaryStart = new Date();
-        summaryStart.setDate(summaryStart.getDate() - 14); 
+        // // instantiate start and end date (past 2 weeks) for summary stats
+        // const summaryStart = new Date();
+        // summaryStart.setDate(summaryStart.getDate() - 14); 
 
-        // add scenario to severity list
-        const sevList = _.cloneDeep(this.state.severityList);
-        sevList[0].scenario = scenario.key;
+        // // add scenario to severity list
+        // const sevList = _.cloneDeep(this.state.severityList);
+        // sevList[0].scenario = scenario.key;
 
         this.setState({
-            dataset,
-            SCENARIOS,
-            scenario,
-            scenarioList,
-            dates: newDates,
-            allTimeDates,
-            seriesList: [filteredSeries],
-            severityList: sevList,
-            allTimeSeries,
-            seriesMax,
-            seriesMin,
-            statThreshold,
-            yAxisLabel,
-            firstDate,
-            lastDate,
-            percExceedenceList,
-            confBoundsList: [filteredConfBounds],
+            dataset
         }, () => {
             this.setState({
                 dataLoaded: true
@@ -156,94 +100,89 @@ class MainContainer extends Component {
         })
     };
 
-    componentDidUpdate(prevProp, prevState) {
-        if (this.state.stat !== prevState.stat ||
-            this.state.scenarioList !== prevState.scenarioList ||
-            this.state.severityList !== prevState.severityList ||
-            this.state.dateRange !== prevState.dateRange ||
-            this.state.r0 !== prevState.r0 ||
-            this.state.dataset !== prevState.dataset) {
+    // componentDidUpdate(prevProp, prevState) {
+    //     if (this.state.dataset !== prevState.dataset) {
 
-            const filteredSeriesList = []
-            const percExceedenceList = []
-            const confBoundsList = [];
-            let brushSeries
+    //         const filteredSeriesList = []
+    //         const percExceedenceList = []
+    //         const confBoundsList = [];
+    //         let brushSeries
             
-            const { dataset, stat, severityList, scenarioList, r0 } = this.state;
-            // filter series and dates by dateRange
-            const idxMin = timeDay.count(this.state.firstDate, this.state.dateRange[0]);
-            const idxMax = timeDay.count(this.state.firstDate, this.state.dateRange[1]);
-            const filteredDates = Array.from(this.state.allTimeDates.slice(idxMin, idxMax));
-            const dateThresholdIdx = Math.ceil(filteredDates.length / 2)
-            const dateThreshold = filteredDates[dateThresholdIdx]
-            let statThreshold = 0
-            let sliderMin = 100000000000
-            let sliderMax = 0
+    //         const { dataset, stat, severityList, scenarioList, r0 } = this.state;
+    //         // filter series and dates by dateRange
+    //         const idxMin = timeDay.count(this.state.firstDate, this.state.dateRange[0]);
+    //         const idxMax = timeDay.count(this.state.firstDate, this.state.dateRange[1]);
+    //         const filteredDates = Array.from(this.state.allTimeDates.slice(idxMin, idxMax));
+    //         const dateThresholdIdx = Math.ceil(filteredDates.length / 2)
+    //         const dateThreshold = filteredDates[dateThresholdIdx]
+    //         let statThreshold = 0
+    //         let sliderMin = 100000000000
+    //         let sliderMax = 0
 
-            for (let i = 0; i < scenarioList.length; i++) {
-                const copy = Array.from(
-                    dataset[scenarioList[i].key][severityList[i].key][stat.key].sims
-                    );
+    //         for (let i = 0; i < scenarioList.length; i++) {
+    //             const copy = Array.from(
+    //                 dataset[scenarioList[i].key][severityList[i].key][stat.key].sims
+    //                 );
 
-                // filter down sims on reproductive number
-                const newSeries = copy.filter(s => {
-                    return (s.r0 > r0[0] && s.r0 < r0[1])
-                });
+    //             // filter down sims on reproductive number
+    //             const newSeries = copy.filter(s => {
+    //                 return (s.r0 > r0[0] && s.r0 < r0[1])
+    //             });
                     
-                const filteredSeriesForStatThreshold = newSeries.map( s => {
-                    const newS = {...s}
-                    newS.vals = s.vals.slice(idxMin, idxMax)
-                    return newS
-                });
+    //             const filteredSeriesForStatThreshold = newSeries.map( s => {
+    //                 const newS = {...s}
+    //                 newS.vals = s.vals.slice(idxMin, idxMax)
+    //                 return newS
+    //             });
 
-                const seriesPeaks = filteredSeriesForStatThreshold.map(sim => sim.max);
-                const [seriesMin, seriesMax] = getRange(seriesPeaks);
-                if (seriesMin < sliderMin) sliderMin = seriesMin
-                if (seriesMax > sliderMax) sliderMax = seriesMax
-                // update dateThreshold before updating statThreshold?
-                if (i === 0) statThreshold = Math.ceil(seriesMax / 1.2);
+    //             const seriesPeaks = filteredSeriesForStatThreshold.map(sim => sim.max);
+    //             const [seriesMin, seriesMax] = getRange(seriesPeaks);
+    //             if (seriesMin < sliderMin) sliderMin = seriesMin
+    //             if (seriesMax > sliderMax) sliderMax = seriesMax
+    //             // update dateThreshold before updating statThreshold?
+    //             if (i === 0) statThreshold = Math.ceil(seriesMax / 1.2);
 
-                const simsOver = this.updateThresholdIterate(
-                    newSeries,
-                    statThreshold,
-                    this.state.allTimeDates,
-                    dateThreshold
-                )
-                if (i === 0) brushSeries = newSeries
+    //             const simsOver = this.updateThresholdIterate(
+    //                 newSeries,
+    //                 statThreshold,
+    //                 this.state.allTimeDates,
+    //                 dateThreshold
+    //             )
+    //             if (i === 0) brushSeries = newSeries
 
-                const percExceedence = simsOver / newSeries.length;
-                percExceedenceList.push(percExceedence)
+    //             const percExceedence = simsOver / newSeries.length;
+    //             percExceedenceList.push(percExceedence)
 
-                const filteredSeries = newSeries.map( s => {
-                    const newS = {...s}
-                    newS.vals = s.vals.slice(idxMin, idxMax)
-                    return newS
-                })
-                filteredSeriesList.push(filteredSeries)
+    //             const filteredSeries = newSeries.map( s => {
+    //                 const newS = {...s}
+    //                 newS.vals = s.vals.slice(idxMin, idxMax)
+    //                 return newS
+    //             })
+    //             filteredSeriesList.push(filteredSeries)
                 
-                // build confidence bounds list
-                const confBounds = dataset[scenarioList[i].key][severityList[i].key][stat.key].conf;
+    //             // build confidence bounds list
+    //             const confBounds = dataset[scenarioList[i].key][severityList[i].key][stat.key].conf;
 
-                // make sure the stat has confidence bounds array
-                if (confBounds && confBounds.length > 0) {
-                    // filter by date range selected
-                    const filteredConfBounds = confBounds.slice(idxMin, idxMax)
-                    confBoundsList.push(filteredConfBounds);
-                }
-            }
-            this.setState({
-                seriesList: filteredSeriesList,
-                allTimeSeries: brushSeries,
-                dates: filteredDates,
-                statThreshold,
-                dateThreshold,
-                seriesMin: sliderMin,
-                seriesMax: sliderMax,
-                percExceedenceList,
-                confBoundsList
-            })
-        }
-    };
+    //             // make sure the stat has confidence bounds array
+    //             if (confBounds && confBounds.length > 0) {
+    //                 // filter by date range selected
+    //                 const filteredConfBounds = confBounds.slice(idxMin, idxMax)
+    //                 confBoundsList.push(filteredConfBounds);
+    //             }
+    //         }
+    //         this.setState({
+    //             seriesList: filteredSeriesList,
+    //             allTimeSeries: brushSeries,
+    //             dates: filteredDates,
+    //             statThreshold,
+    //             dateThreshold,
+    //             seriesMin: sliderMin,
+    //             seriesMax: sliderMax,
+    //             percExceedenceList,
+    //             confBoundsList
+    //         })
+    //     }
+    // };
 
     componentWillUnmount() {
         window.removeEventListener('resize', this.updateGraphDimensions)
@@ -262,288 +201,70 @@ class MainContainer extends Component {
         this.setState({ mapContainerW, mapContainerH });
     }
 
-    updateThresholdIterate = (series, statThreshold, dates, dateThreshold) => {
-        const dateIndex = dates.findIndex(
-            date => formatDate(date) === formatDate(dateThreshold)
-            );
-        let simsOver = 0;
-        Object.values(series).forEach((sim, simIdx) => {
-            let simOver = false;
-            for (let i = 0; i < dateIndex; i++) {
-                if (sim.vals[i] > statThreshold){
-                    simsOver = simsOver + 1;
-                    simOver = true;
-                    break;
-                }
-            }
-            simOver ? sim.over = true : sim.over = false
-        })
-        return simsOver;
-    }
+    // initializeGeoid(geoid, dataset) {
+    //     // const [SCENARIOS, scenario, scenarioList, scenarioListChart, scenarioMap] =
+    //     //     instantiateScenarios(dataset);
 
-    updateThreshold = (series, statThreshold, dates, dateThreshold) => {
-        // update 'over' flag to true if sim peak surpasses statThreshold
-        // returns numSims 'over' threshold
-        // first find index of dates at dateThreshold
-        const dateIndex = dates.indexOf(dateThreshold);
+    //     // re-initialize severity
+    //     // const severityList = [_.cloneDeep(LEVELS[0])];
+    //     // severityList[0].scenario = scenario.key;
 
-        let simsOver = 0;
-        Object.values(series).map(sim => {
-            // calculate max once, use to find maxIndex
-            const maxIdx = maxIndex(sim.vals)
-            const dateAtMax = dates[maxIdx]
-            // we need to keep track of whether simval at dateThreshold is over statThreshold
-            // as well as whether the max is over statThreshold and occured in the past
-            if (sim.vals[dateIndex] > statThreshold
-                || (dateAtMax < dateThreshold && sim.max > statThreshold)) {
-                simsOver = simsOver + 1;
-                return sim.over = true;
-            } else {
-                return sim.over = false;
-            }
-        })
-        return simsOver;
-    }
+    //     // re-initialize countyBoundaries
+    //     // const state = geoid.slice(0, 2);
+    //     // const countyBoundaries = require('../store/countyBoundaries.json')[state];
+    //     // const statsForCounty = geojsonStats[state];
 
-    initializeGeoid(geoid, dataset) {
-        const [SCENARIOS, scenario, scenarioList, scenarioListChart, scenarioMap] =
-            instantiateScenarios(dataset);
-
-        // re-initialize severity
-        const severityList = [_.cloneDeep(LEVELS[0])];
-        severityList[0].scenario = scenario.key;
-
-        // re-initialize countyBoundaries
-        const state = geoid.slice(0, 2);
-        const countyBoundaries = require('../store/countyBoundaries.json')[state];
-        const statsForCounty = geojsonStats[state];
-
-        this.setState({
-            dataset,
-            geoid,
-            SCENARIOS,
-            scenarioList,
-            scenarioListChart,
-            scenarioMap,
-            severityList,
-            countyBoundaries,
-            statsForCounty,
-            r0: [0, 4]
-        })
-    }
+    //     this.setState({
+    //         dataset,
+    //         geoid,
+    //         SCENARIOS,
+    //         scenarioList,
+    //         scenarioListChart,
+    //         scenarioMap,
+    //         severityList,
+    //         countyBoundaries,
+    //         statsForCounty,
+    //         r0: [0, 4]
+    //     })
+    // }
 
     handleCountySelect = (i) => {
         const dataset = require(`../store/geo${i.geoid}.json`);
-        this.initializeGeoid(i.geoid, dataset);
-    }
+        // this.initializeGeoid(i.geoid, dataset);
+
+        this.setState({
+            dataset,
+            geoid: i.geoid
+        })
+    };
     
     handleUpload = (dataset, geoid) => {
-        this.initializeGeoid(geoid, dataset);
-    };
-
-    handleButtonClick = (i) => {
-        const yAxisLabel = `Daily ${i.name}`;
+        // this.initializeGeoid(geoid, dataset);
         this.setState({
-            stat: i, 
-            yAxisLabel,
-            r0: [0, 4]
+            dataset,
+            geoid
         })
     };
-
-    handleScenarioClickGraph = (items) => {
-        // items is Array of scenario names
-        const scenarioClkCntr = this.state.scenarioClickCounter + 1;
-        let newScenarios = [];
-        let newSevs = [];
-
-        for (let item of items) {
-            const defaultSev = _.cloneDeep(LEVELS[0]); 
-            defaultSev.scenario = item;
-            newSevs.push(defaultSev)
-
-            const scenario = this.state.SCENARIOS.filter(s => s.key === item)[0];
-            newScenarios.push(scenario);
-        }
-
-        this.setState({
-            scenarioList: newScenarios,
-            scenarioClickCounter: scenarioClkCntr,
-            severityList: newSevs, 
-            r0: [0, 4]
-        })        
-    };
-
-    handleSeveritiesClick = (i) => {
-        let newSevList = _.cloneDeep(this.state.severityList);
-        newSevList.forEach(sev => {
-            if (sev.scenario === i.scenario) {
-                return sev.key = i.key;
-            }
-        })
-        this.setState({
-            severityList: newSevList, 
-            r0: [0, 4],
-            animateTransition: true
-        });
-    };
-
-    handleSeveritiesHover = (i) => {this.setState({scenarioHovered: i})};
-
-    handleSeveritiesHoverLeave = () => {this.setState({scenarioHovered: ''});}
-
-    handleR0Change = (e) => {
-        this.setState({ r0: e, animateTransition: false })
-    };
-
-    handleStatSliderChange = (thresh) => {
-        const { dates, dateThreshold, allTimeDates } = this.state;
-        // const rounded = Math.ceil(i / 100) * 100;
-        const copyList = Array.from(this.state.seriesList);
-        const allSeriesCopy = Array.from(this.state.allTimeSeries);
-        this.updateThresholdIterate(allSeriesCopy, thresh, allTimeDates, dateThreshold);
-        const percExceedenceList = [];
-
-        for (let i = 0; i < copyList.length; i++) {
-            const simsOver = this.updateThresholdIterate(copyList[i], thresh, dates, dateThreshold);
-            const percExceedence = simsOver / copyList[i].length;
-            percExceedenceList.push(percExceedence);
-        }
-
-        this.setState({
-            seriesList: copyList,
-            allTimeSeries: allSeriesCopy,
-            statThreshold: +thresh,
-            percExceedenceList,
-            animateTransition: true
-        });
-    };
-
-    handleDateSliderChange = (thresh) => {
-        const { statThreshold, dates, allTimeDates } = this.state;
-        const copyList = Array.from(this.state.seriesList);
-        const allSeriesCopy = Array.from(this.state.allTimeSeries);
-        this.updateThresholdIterate(allSeriesCopy, statThreshold, allTimeDates, thresh);
-        const percExceedenceList = [];
-
-        for (let i = 0; i < copyList.length; i++) {
-            const simsOver = this.updateThresholdIterate(copyList[i], statThreshold, dates, thresh);
-            const percExceedence = simsOver / copyList[i].length;
-            percExceedenceList.push(percExceedence);
-        }
-        this.setState({
-            seriesList: copyList,
-            allTimeSeries: allSeriesCopy,
-            dateThreshold: thresh,
-            percExceedenceList
-        })
-    }
-
-    handleBrushRange = (i) => {this.setState({dateRange: i, animateTransition: false});};
-
-    handleBrushStart = () => {this.setState({brushActive: true, animateTransition: false})}
-
-    handleBrushEnd = () => {this.setState({brushActive: false, animateTransition: true})}
-
-    // handleScaleToggle = (scale) => {this.setState({ summaryScale: scale })}
-
-    handleConfClick = () => {
-        this.setState(prevState => ({showConfBounds: !prevState.showConfBounds}));
-    };
-
-    handleActualClick = () => {
-        this.setState(prevState => ({showActual: !prevState.showActual}));
-    };
-
-    handleSliderMouseEvent = (type, slider, view) => {
-        if (view === 'graph') {
-            if (slider === 'stat') {
-                if (type === 'mousedown') {
-                    this.setState({ statSliderActive: true })
-                } else {
-                    this.setState({ statSliderActive: false })
-                }
-            } else {
-                if (type === 'mousedown') {
-                    this.setState({ dateSliderActive: true })
-                } else {
-                    this.setState({ dateSliderActive: false })
-                }
-            }
-        } 
-    }
-
-    toggleAnimateTransition = () => {
-        this.setState({ animateTransition: !this.state.animateTransition })
-    }
 
     render() {
         return (
             <Layout>
                 {/* Search Component */}
                 <Search
-                    stat={this.state.stat}
                     geoid={this.state.geoid}
+                    stat={this.state.stat} // TODO: do you need this?
                     onFileUpload={this.handleUpload}
                     onCountySelect={this.handleCountySelect}>
                 </Search>
 
                 {/* MainGraph Component */}
                 {this.state.dataLoaded &&
-                <div>
-                    <MainGraph 
-                        stat={this.state.stat}
-                        geoid={this.state.geoid}
-                        yAxisLabel={this.state.yAxisLabel}
-                        scenarioList={this.state.scenarioList}
-                        severity={this.state.severity}
-                        r0={this.state.r0}
-                        animateTransition={this.state.animateTransition}
-                        toggleAnimateTransition={this.toggleAnimateTransition}
-                        simNum={this.state.simNum}
-                        showConfBounds={this.state.showConfBounds}
-                        confBoundsList={this.state.confBoundsList}
-                        showActual={this.state.showActual}
-                        seriesList={this.state.seriesList}
-                        dates={this.state.dates}
-                        statThreshold={this.state.statThreshold}
-                        dateThreshold={this.state.dateThreshold}
-                        percExceedenceList={this.state.percExceedenceList}
-                        dateRange={this.state.dateRange}
-                        brushActive={this.state.brushActive}
-                        width={this.state.graphW}
-                        height={this.state.graphH}
-                        scenarioClickCounter={this.state.scenarioClickCounter}
-                        scenarioHovered={this.state.scenarioHovered}
-                        statSliderActive={this.state.statSliderActive}
-                        dateSliderActive={this.state.dateSliderActive}
-                        allTimeSeries={this.state.allTimeSeries}
-                        allTimeDates={this.state.allTimeDates}
-                        x={margin.yAxis}
-                        y={0}
-                        onBrushChange={this.handleBrushRange}
-                        onBrushStart={this.handleBrushStart}
-                        onBrushEnd={this.handleBrushEnd}
-                        SCENARIOS={this.state.SCENARIOS}
-                        scenario={this.state.scenario}
-                        onScenarioClickGraph={this.handleScenarioClickGraph}
-                        onButtonClick={this.handleButtonClick}
-                        onConfClick={this.handleConfClick}
-                        severityList={this.state.severityList}
-                        onSeveritiesClick={this.handleSeveritiesClick}
-                        onSeveritiesHover={this.handleSeveritiesHover}
-                        onSeveritiesHoverLeave={this.handleSeveritiesHoverLeave}
-                        onHandleR0Change={this.handleR0Change}
-                        seriesMax={this.state.seriesMax}
-                        seriesMin={this.state.seriesMin}
-                        dateThresholdIdx={this.state.dateThresholdIdx}
-                        firstDate={this.state.firstDate}
-                        lastDate={this.state.lastDate}
-                        onStatSliderChange={this.handleStatSliderChange}
-                        onDateSliderChange={this.handleDateSliderChange}
-                        onSliderMouseEvent={this.handleSliderMouseEvent}
-                    />
-                </div>
-                }
+                <MainGraph 
+                    geoid={this.state.geoid}
+                    dataset={this.state.dataset}
+                    width={this.state.graphW}
+                    height={this.state.graphH}
+                />}
 
                 {/* MainChart Component */}
                 {this.state.dataLoaded &&
